requirements.txt
flask
flask-sqlalchemy
gunicorn
numpy
pandas
Werkzeug
fpdf
openpyxl
qrcode[pil]
pyinstaller